+++ src/main/java/webserver/controller/FrontController.java
@@ -3,7 +3,7 @@ package webserver.controller;
 import webserver.controller.annotation.Handler;
 import webserver.controller.support.HandlerMethod;
 import webserver.controller.support.PathMapKey;
-import webserver.controller.support.RequestController;
+import webserver.controller.annotation.RequestController;
 import webserver.request.Request;
 
 import java.lang.reflect.Method;
