+++ src/main/java/webserver/controller/HomeController.java
@@ -3,7 +3,7 @@ package webserver.controller;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
 import webserver.controller.annotation.Handler;
-import webserver.controller.support.RequestController;
+import webserver.controller.annotation.RequestController;
 import webserver.request.Request;
 import webserver.response.Response;
