+++ src/main/java/webserver/controller/UserController.java
@@ -5,7 +5,7 @@ import model.User;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.HttpStatus;
 import webserver.controller.annotation.Handler;
-import webserver.controller.support.RequestController;
+import webserver.controller.annotation.RequestController;
 import webserver.request.Request;
 import webserver.request.support.FormData;
 import webserver.request.support.QueryParameters;
